requirements.txt
matplotlib
miditoolkit==0.1.16
mido==1.2.10
numpy==1.22.4
ortools==9.5.2237
pandas==1.5.3
parmap==1.5.3
pretty_midi==0.2.10
pydantic==1.9.1
PyYAML==6.0
torch
tqdm==4.64.1
yacs==0.1.8
Flask
